refactor: itemsCardコンポーネントをリファクタリング
 - 商品カードコンポーネントを新規作成
 - itemsCardから商品カード部分の記述を切り出して、商品カードコンポーネントに移動させました
